% Generated by roxygen2 (4.0.1): do not edit by hand
\name{cross}
\alias{cross}
\title{Cross product of vectors.}
\usage{
cross(a, b, unit = FALSE)
}
\arguments{
\item{a}{numeric 3-vector}

\item{b}{numeric 3-vector}

\item{unit}{logical; return a unit vector?}
}
\value{
3-vector giving the standard vector cross product, or a unit
vector in the same direction, if \code{unit} is TRUE.
The cross product has magnitude \code{|a||b|sin(theta)} where \code{theta}
is the (smaller) angle between the two vectors.  The direction is perpendicular
to both vectors, with sign chosen by the right hand rule.
}
\description{
Computer the Euclidean cross product of two 3-vectors.
}
\author{
John Brzustowski \email{jbrzusto@REMOVE_THIS_PART_fastmail.fm}
}

