% Generated by roxygen2 (4.0.1): do not edit by hand
\name{dot}
\alias{dot}
\title{Dot product of vectors.}
\usage{
dot(x1, x2)
}
\arguments{
\item{x1}{left vector, either 3-vector or \code{n x 3} matrix}

\item{x2}{right vector, either 3-vector or \code{n x 3} matrix}
}
\value{
vector or matrix of dot products.  If both \code{x1} and \code{x2}
are 3-vectors, then the result is a scalar.  If \code{x1} is a 3-vector
but \code{x2} is an \code{n x 3} matrix, the results is a vector of length
\code{n} giving the dot product of x1 with each row of x2.  Similarly
if \code{x1} is an \code{n x 3} matrix and \code{x2} is a 3-vector.
Finally, if \code{x1} is \code{m x 3} and \code{x2} is  \code{m x 3},
returns an \code{m x n} matrix with entry \code{[i, j]} giving the
dot product of the \code{i}-th row of \code{x1} with the \code{j}-th row
of \code{x2}.
}
\description{
Computer the (real) Euclidean dot product of 3-vectors.
}
\author{
John Brzustowski \email{jbrzusto@REMOVE_THIS_PART_fastmail.fm}
}

