% Generated by roxygen2 (4.0.1): do not edit by hand
\name{gain.omni}
\alias{gain.omni}
\title{Gain of an omnidirectional antenna at remote points.}
\usage{
gain.omni(antPos, antAxis, remPos)
}
\arguments{
\item{antPos}{3-element antenna position vector (x, y, z)}

\item{antAxis}{3-element antenna orientation axis (x, y, z). The
conducting element of the antenna is parallel to this vector.}

\item{remPos}{3-element vector (x, y, z) or n x 3 matrix of remote locations
at which gain is to be calculated.}
}
\value{
gain factor, in linear units, with maximum possible value of 1.0
}
\description{
Given the position and orientation of an omnidirectional antenna,
calculate the one-way power gain at remote points.
}
\note{
This gain factor is purely due to far-field antenna pattern, and does not
consider distance to the remote point(s) or relative polarization of
remote antenna.
}
\author{
John Brzustowski \email{jbrzusto@REMOVE_THIS_PART_fastmail.fm}
}
\references{
\url{http://en.wikipedia.org/wiki/Dipole_antenna#Half-wave_dipole}
}

