% Generated by roxygen2 (4.0.1): do not edit by hand
\name{getOperatingStatus}
\alias{getOperatingStatus}
\title{Get receiver operating status via hourly file presence.}
\usage{
getOperatingStatus(site, proj, year = NULL)
}
\arguments{
\item{site}{site name}

\item{proj}{project name}

\item{year}{deployment year (defaults to current year)}
}
\value{
a data frame with columns \code{ts}, giving the top-of-the-hour timestamp
for each hour of each day in the full operating range, and \code{n}, the count of files
written within that hour.
}
\description{
Counts the number of files written by a receiver during each hour for
which data are available.
}
\author{
John Brzustowski \email{jbrzusto@REMOVE_THIS_PART_fastmail.fm}
}

