% Generated by roxygen2 (4.0.1): do not edit by hand
\name{makeAxis}
\alias{makeAxis}
\title{Generate a 3-D unit axis vector.}
\usage{
makeAxis(dir, ref = NULL)
}
\arguments{
\item{dir}{direction of antenna axis, which can be specified several ways:

\enumerate{
\item "horizontal": the elements are parallel to the ground and perpendicular to the
reference axis.
\item "vertical": the elements are vertical

\item a single number: if \code{ref} is specified, then \code{dir}
is the angle of rotation of the elements when viewed from the
antenna base (wider end) toward the antenna tip.  Angle is measured
clockwise in degrees, starting at 0 for horizontal.  So 0 is the
same as "horizontal" and 90 is the same as "vertical".

If \code{ref} is \code{NULL}, then \code{dir} is the azimuth angle
(degrees clockwise from North) of an axis parallel to the ground.

\item a 2-vector: this represents azimuth and elevation angles (degrees clockwise
from North, degrees above the horizon) of the axis.
\item a 3-vector: this is a vector giving the axis as a direction vector (x, y, z).
}}

\item{ref:}{reference axis.  If not null, must be a unit 3-vector.}
}
\value{
a unit 3-vector giving the actual direction of the axis
}
\description{
Creates a unit 3-vector indicating an axis direction.  Various ways of specifying
the direction in 3D are allowed.
}
\author{
John Brzustowski \email{jbrzusto@REMOVE_THIS_PART_fastmail.fm}
}

