% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeOmni.R
\name{makeOmni}
\alias{makeOmni}
\title{Create an omnidirectional antenna object.}
\usage{
makeOmni(pos, axis)
}
\arguments{
\item{pos}{3-element antenna position vector (x, y, z)}

\item{axis}{3-element antenna orientation axis (x, y, z). This
points along the antenna.  See \code{\link{makeAxis}}, for other forms
in which you can specify this parameter.}
}
\description{
Creates a 'proto' object of class 'antenna' which will store
position and orientation of an omnidirectional antenna.
}
\author{
John Brzustowski \email{jbrzusto@REMOVE_THIS_PART_fastmail.fm}
}
\references{
\url{http://en.wikipedia.org/wiki/Dipole_antenna#Half-wave_dipole}
}

