% Generated by roxygen2 (4.0.1): do not edit by hand
\name{makeYagi}
\alias{makeYagi}
\title{Create a yagi antenna object.}
\usage{
makeYagi(n, pos, axis1, axis2)
}
\arguments{
\item{n}{integer; number of elements (cross pieces) of the antenna.
So far, only antennas with 5 and 9 elements are supported.}

\item{pos}{3-element antenna position vector (x, y, z)}

\item{axis1}{orientation of main antenna axis.  This is one of the following:
\enumerate{
  \item "vertical"; this represents an atenna pointing straight up
  \item a single number; this represents the azimuth (degrees clockwise from North)
of an antenna oriented parallel to the ground
  \item a 2-vector; this represents azimuth and elevation angles (degrees clockwise
from North, degrees above the horizon) for an antenna pointing in an arbitrary direction
  \item a 3-vector; this is a vector giving the direction (x, y, z) in which the antenna
is pointing
}}

\item{axis2}{orientation of the antenna elements ('cross pieces') relative to the antenna axis.
Specified as one of the following:

\enumerate{
\item "horizontal": the elements are parallel to the ground
\item "vertical": the elements are vertical
\item a single number: the angle of rotation of the elements when viewed from the antenna
base (wider end) toward the antenna tip.  Angle is measured clockwise in degrees, starting
at 0 for horizontal.  So 0 is the same as "horizontal" and 90 is the same as "vertical".
\item a 2-vector: this represents azimuth and elevation angles (degrees clockwise
from North, degrees above the horizon) for the antenna elements, and must be perpendicular to
the \code{axis1} parameter
\item a 3-vector: this is a vector giving the direction (x, y, z) in which the antenna elements
are pointing.  It must be perpendicular to the \code{axis1} parameter.
}}
}
\description{
Creates a 'proto' object of class 'antenna' which will store
position and axis orientations of a Yagi-Uda antenna.
}
\author{
John Brzustowski \email{jbrzusto@REMOVE_THIS_PART_fastmail.fm}
}

