% Generated by roxygen2 (4.0.1): do not edit by hand
\name{plotOperatingStatus}
\alias{plotOperatingStatus}
\title{Plot receiver operating status.}
\usage{
plotOperatingStatus(site, proj, year = lubridate::year(Sys.time()))
}
\arguments{
\item{site}{site name}

\item{proj}{project name}

\item{year}{deployment year (defaults to current year)}
}
\value{
nothing

Plots to the current graphics device.
}
\description{
Plots a summary of the status data generated by \code{getOperatingStatus()}
}
\author{
John Brzustowski \email{jbrzusto@REMOVE_THIS_PART_fastmail.fm}
}

