% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/publicLotekToSG.R
\name{publicLotekToSG}
\alias{publicLotekToSG}
\title{filter and convert public Lotek tag detections to SG format}
\usage{
publicLotekToSG(absDir, year, tagDB, ...)
}
\arguments{
\item{absDir}{An absolute directory in the SensorGnome hierarchy which
corresponds to a single site. All files in the folder with names
ending in "_raw_public_tags.rds" should be from the same receiver,
which is treated as a site.  The project name and site name are
taken to be the last two components of the path.
e.g. /SG/contrib/2013/taylor/sonora}

\item{year}{the deployment year}

\item{tagDB}{a .csv file of tag registrations, containing only the
public fields proj, id, freq, bi}

\item{...}{further parameters to filterTags(), which is called on
the merged dataset.}
}
\value{
A dataframe compatible with SensorGnome databases, consisting
of non-duplicated filtered tag records present in \code{tagDB}

As a side-effect, the dataframe is written to a file called
YEAR_PROJ_SITE_alltags.rds in the folder \code{path}.
}
\description{
A set of .rds files generated by publicDTA() from a set of Lotek
DTA files for a single receiver are merged, filtered, and output
as a SensorGnome-style site tag dataset.
}
\author{
John Brzustowski \email{jbrzusto is at fastmail.fm}
}

