% Generated by roxygen2 (4.0.1): do not edit by hand
\name{SGify}
\alias{SGify}
\title{Convert a data.frame of Lotek tag data to a SensorGnome-compatible data.frame}
\usage{
SGify(tags, year, proj, site, recv)
}
\arguments{
\item{tags}{a data.frame of Lotek tag data, as produced by readDTA() or filterDTA()}

\item{year}{the deployment year.}

\item{recv}{the receiver model / serial number (character scalar)}
}
\value{
A data.frame with the same number of rows, but in SensorGnome format
}
\description{
The output of filterDTA() is augmented with additional columns required by the SensorGnome
database format.
}
\author{
John Brzustowski \email{jbrzusto is at fastmail.fm}
}

