% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/windowApply.R
\name{windowApply}
\alias{windowApply}
\title{Apply a function to one variable on each sliding window of fixed width
in another variable.}
\usage{
windowApply(x, y, w, f)
}
\arguments{
\item{x}{- vector of values in non-decreasing order}

\item{y}{- vector of arbitrary values, with
\code{length(x)==length(y)}}

\item{w}{- width of sliding window in same units as \code{x}}

\item{f}{- function of one or two variables to apply to each
sliding window of values.  \code{f} is called on each window of data,
with windowed y values if \code{f} accepts one argument and windowed
x and y values if \code{f} takes two arguments.}
}
\value{
result of applying \code{f} to a sliding window of y
values; this has the same length as \code{x} and \code{y}.  The
\code{i}th element is obtained by applying \code{f} to
\code{y[i:j]} where \code{j} is the largest index such that \code{j
<= length(x)} and {x[j]-x[i] < w}
}
\description{
Given a set of (x,y) pairs with x in non-decreasing order, apply a
function to each maximal set of consecutive pairs where the range
of x values is at most a fixed value w, the window size.  For
example, if x is time, y is temperature, w is 30 and the function
is \code{mean}, calculate the mean temperature over a sliding
window of width 30 seconds.
}
\note{
The number of elements of y used to compute each element in
the result will not in general be constant.  It will be constant if
the values of x are equally spaced, except that the tail end of the
result will be computed with fewer elements.
}
\author{
John Brzustowski \email{jbrzusto@REMOVE_THIS_PART_fastmail.fm}
}

