% Generated by roxygen2 (4.0.1): do not edit by hand
\name{friis}
\alias{friis}
\title{Friis power gain between two antennas.}
\usage{
friis(ant1, ant2, waveLength = 3e+08/166380000)
}
\arguments{
\item{ant1}{first antenna (a 'proto' object of class 'antenna')}

\item{ant2}{second antenna (a 'proto' object of class 'antenna')}

\item{waveLength}{in metres.  Defaults to the wavelength of the
telemetry band frequency 166.38 MHz}
}
\description{
Given the position and orientation of two antennas, calculate the
one-way free space power gain between them, taking into account
distance, antenna patterns, and polarization, but not atmospheric
effects.
}
\author{
John Brzustowski \email{jbrzusto@REMOVE_THIS_PART_fastmail.fm}
}
\references{
\url{http://en.wikipedia.org/wiki/Dipole_antenna#Half-wave_dipole}
}

