% Generated by roxygen2 (4.0.1): do not edit by hand
\name{getOperatingStatus}
\alias{getOperatingStatus}
\title{Get receiver operating status.}
\usage{
getOperatingStatus(site, proj, year = lubridate::year(Sys.time()))
}
\arguments{
\item{site}{site name}

\item{proj}{project name}

\item{year}{deployment year (defaults to current year)}
}
\value{
a list with these items:\cr
\describe{
\item{files.per.hour}{data frame with at most one record per hour, giving number of raw data files recorded and bootcount}
\item{gpx.fix}{data frame with at most hourly true GPS fix (i.e. repeated 'stuck' fixes are not reported)}
\item{undated.period}{the number of days of data recorded with dates beginning with 1 Jan 2000; these are periods when the GPS had not set the clock}
\item{num.boots}{the number of times the system was restarted during the operating period}
}
}
\description{
Counts the number of files written by a receiver during each hour for
which data are available.  Examines GPS records to see when GPS was
functioning (vs. stuck on a fix). For each day, finds all antennas
for which at least one pulse was detected.
}
\author{
John Brzustowski \email{jbrzusto@REMOVE_THIS_PART_fastmail.fm}
}

