% Generated by roxygen2 (4.0.1): do not edit by hand
\name{mag}
\alias{mag}
\title{Magnitudes of vectors.}
\usage{
mag(x)
}
\arguments{
\item{x}{vector or \code{n x 3} matrix}
}
\value{
the norm of x, or an n-vector giving the norm of each row of x (for a matrix)
}
\description{
Get the Euclidean norm of a single vector or each row of an nx3 matrix.
}
\author{
John Brzustowski \email{jbrzusto@REMOVE_THIS_PART_fastmail.fm}
}

