% Generated by roxygen2 (4.0.1): do not edit by hand
\name{makeOmni}
\alias{makeOmni}
\title{Create an omnidirectional antenna object.}
\usage{
makeOmni(pos, axis)
}
\arguments{
\item{pos}{3-element antenna position vector (x, y, z)}

\item{axis}{3-element antenna orientation axis (x, y, z). This
points along the antenna.}
}
\description{
Creates a 'proto' object of class 'antenna' which will store
position and orientation of an omnidirectional antenna.
}
\author{
John Brzustowski \email{jbrzusto@REMOVE_THIS_PART_fastmail.fm}
}
\references{
\url{http://en.wikipedia.org/wiki/Dipole_antenna#Half-wave_dipole}
}

