% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotOperatingStatus.R
\name{plotOperatingStatus}
\alias{plotOperatingStatus}
\title{Plot receiver operating status.}
\usage{
plotOperatingStatus(site, proj, year = lubridate::year(Sys.time()), pngFile)
}
\arguments{
\item{site}{site name}

\item{proj}{project name}

\item{year}{deployment year (defaults to current year)}

\item{pngFile}{full path to output .png filename}
}
\value{
a data frame with these columns
\describe{
\item{files.per.hour}{data frame with at most one record per hour, giving number of raw data files recorded and bootcount}
\item{gps.fix}{data frame with at most hourly true GPS fix (i.e. repeated 'stuck' fixes are not reported)}
\item{undated.period}{the number of days of data recorded with dates beginning with 1 Jan 2000; these are periods when the GPS had not set the clock}
\item{num.boots}{the number of times the system was restarted during the operating period}
\item{pulse.counts}{the number of pulses per antenna for each hour}
}

Plots to the current graphics device.
}
\description{
Plots a summary of the status data generated by \code{getOperatingStatus()}
}
\author{
John Brzustowski \email{jbrzusto@REMOVE_THIS_PART_fastmail.fm}
}

